Shell Improvements

+++ TPE_ARQUI/Userland/SampleCodeModule/include/test_prio.h
@@ -2,6 +2,6 @@
 #define _TEST_PRIO
 
 #include <stdint.h>
-int64_t test_prio(void);
+int test_prio(void);
 
 #endif
