+++ TPE_ARQUI/Userland/SampleCodeModule/tests/test_mm.c
@@ -16,7 +16,7 @@ typedef struct MM_rq {
 	uint32_t size;
 } mm_rq;
 
-uint64_t test_mm(uint64_t argc, char *argv[]) {
+int test_mm(uint64_t argc, char *argv[]) {
 	printStr("Testing Memory...\n");
 	mm_rq mm_rqs[MAX_BLOCKS];
 	uint8_t rq;
